結果発表レイアウトの追加とIDの設定
