+++ app/src/main/AndroidManifest.xml
@@ -9,6 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.QuizApp">
+        <activity android:name=".EndActivity">
+            android:parentActivityName="com.example.quizapp.MainActivity">
+            <!-- The meta-data tag is required if you support API level 15 and lower -->
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="com.example.quizapp.MainActivity" />
+        </activity>
         <activity android:name=".JudgeActivity">
             android:parentActivityName="com.example.quizapp.MainActivity">
             <!-- The meta-data tag is required if you support API level 15 and lower -->
